# perf: use libxml2 patch for faster xpath context init