**What problem is this PR intended to solve?**

After chatting with @nwellnhof in #3378, he went ahead and made an performance-focused change upstream in libxml2/gnome@bf5fcf6e that I'm pulling in here to the vendored library.

Benchmark comparing this PR against v1.17.x ("main"):

    Comparison:
             large: main:     3910.6 i/s
          large: patched:     3759.6 i/s - same-ish: difference falls within error

    Comparison:
          small: patched:   242901.7 i/s
             small: main:   127486.0 i/s - 1.91x  slower

We could get greater performance gains by re-using `XPathContext` objects, but only at the cost of a significant amount of additional complexity, since in order to properly support recursive XPath evaluation, Nokogiri would have to push and pop "stack frames" containing:

- internal state contextSize and proximityPosition
- registered namespaces
- registered variables
- function lookup handler

That feels like a lot of code for a small win.

Comparatively, pulling in this upstream patch is still a ~2x speedup with zero additional complexity. That's a win.

**Have you included adequate test coverage?**

N/A


**Does this change affect the behavior of either the C or the Java implementations?**

Performance improvement impacting CRuby only.